+++ src/app/services/collateral.service.ts
@@ -3,15 +3,17 @@ import * as BN from 'bn.js';
 import { ContractsService } from './contracts.service';
 import { Tx, Type } from '../services/tx.service';
 import { Loan } from './../models/loan.model';
+import { Collateral } from './../models/collateral.model';
 import { Utils } from './../utils/utils';
 import { Currency } from './../utils/currencies';
-import { CurrencyItem } from './../services/currencies.service';
+import { CurrenciesService, CurrencyItem } from './../services/currencies.service';
 
 @Injectable()
 export class CollateralService {
 
   constructor(
-    private contractsService: ContractsService
+    private contractsService: ContractsService,
+    private currenciesService: CurrenciesService
   ) { }
 
   /**
@@ -72,6 +74,44 @@ export class CollateralService {
       default:
         return false;
     }
+  }
+
+  /**
+   * Calculate liquidation price
+   * @return Liquidation price in collateral amount
+   */
+  async calculateLiquidationPrice(loan: Loan, collateral: Collateral): Promise<number> {
+    const { liquidationRatio, amount, token } = collateral;
+    const currency: CurrencyItem =
+      this.currenciesService.getCurrencyByKey('address', token.toLowerCase());
+    const liquidationPercentage: string =
+      this.rawToPercentage(liquidationRatio).toString();
+    const collateralPercentage =
+      await this.calculateCollateralPercentage(loan, currency, amount);
+
+    const decimals: number = new Currency(currency.symbol).decimals;
+    const liquidationPrice: number = (Number(liquidationPercentage) * Number(amount)) / Number(collateralPercentage);
+
+    const formattedLiquidationPrice: number =
+      (liquidationPrice as any / 10 ** decimals);
+
+    return formattedLiquidationPrice;
+  }
+
+  /**
+   * Get rate loan currency / collateral currency
+   * @return Exchange rate
+   */
+  async getCollateralRate(loan: Loan, collateral: Collateral): Promise<string> {
+    const loanRate: BN | string =
+      await this.contractsService.getRate(loan.oracle.address, loan.currency.decimals);
+
+    const collateralCurrency = this.currenciesService.getCurrencyByKey('address', collateral.token);
+    const collateralDecimals: number = new Currency(collateralCurrency.symbol).decimals;
+    const collateralRate: BN | string = await this.contractsService.getRate(collateral.oracle, collateralDecimals);
+
+    const rate = Utils.formatAmount((loanRate as any) / (collateralRate as any));
+    return rate;
   }
 
   /**
