+++ src/app/views/loan-detail/detail-collateral/detail-collateral.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit, OnChanges, Input, Output, EventEmitter } from '@angular/core';
 import { MatDialog, MatDialogConfig } from '@angular/material';
-import * as BN from 'bn.js';
 import { DialogCollateralComponent } from '../../../dialogs/dialog-collateral/dialog-collateral.component';
 import { environment } from '../../../../environments/environment';
 // App Models
@@ -105,11 +104,11 @@ export class DetailCollateralComponent implements OnInit, OnChanges {
 
     // set exchange rate
     // TODO: add support for more currencies than rcn
-    const rate = await this.getRate();
+    const rate = await this.collateralService.getCollateralRate(this.loan, this.collateral);
     this.currentExchangeRate = Utils.formatAmount(1 / (rate as any));
 
     // set liquidation price
-    const liquidationPrice = await this.calculateLiquidationPrice();
+    const liquidationPrice = await this.collateralService.calculateLiquidationPrice(this.loan, this.collateral);
     this.currentLiquidationPrice = Utils.formatAmount(1 / (rate as any) * liquidationPrice);
   }
 
@@ -126,48 +125,6 @@ export class DetailCollateralComponent implements OnInit, OnChanges {
       currency,
       amount
     );
-  }
-
-  /**
-   * Calculate liquidation price
-   * @return Liquidation price in collateral amount
-   */
-  async calculateLiquidationPrice(): Promise<number> {
-    const loan: Loan = this.loan;
-    const collateral: Collateral = this.collateral;
-    const { liquidationRatio, amount, token } = collateral;
-    const currency: CurrencyItem =
-      this.currenciesService.getCurrencyByKey('address', token.toLowerCase());
-    const liquidationPercentage: string =
-      this.collateralService.rawToPercentage(liquidationRatio).toString();
-    const collateralPercentage =
-      await this.collateralService.calculateCollateralPercentage(loan, currency, amount);
-
-    const decimals: number = new Currency(currency.symbol).decimals;
-    const liquidationPrice: number = (Number(liquidationPercentage) * Number(amount)) / Number(collateralPercentage);
-
-    const formattedLiquidationPrice: number =
-      (liquidationPrice as any / 10 ** decimals);
-
-    return formattedLiquidationPrice;
-  }
-
-  /**
-   * Get rate loan currency / collateral currency
-   * @return Exchange rate
-   */
-  async getRate(): Promise<string> {
-    const loan: Loan = this.loan;
-    const loanRate: BN | string =
-      await this.contractsService.getRate(loan.oracle.address, loan.currency.decimals);
-
-    const collateral: Collateral = this.collateral;
-    const collateralCurrency = this.currenciesService.getCurrencyByKey('address', collateral.token);
-    const collateralDecimals: number = new Currency(collateralCurrency.symbol).decimals;
-    const collateralRate: BN | string = await this.contractsService.getRate(collateral.oracle, collateralDecimals);
-
-    const rate = Utils.formatAmount((loanRate as any) / (collateralRate as any));
-    return rate;
   }
 
   /**
